Add toolbar in Scripting Window

+++ source/selection.cpp
@@ -365,7 +365,7 @@ wxThread::ExitCode SelectionThread::Entry()
 	Map& map = editor.map;
 
 	if(positions) {
-		for(int i = 0; i < positions->size(); ++i) {
+		for(size_t i = 0; i < positions->size(); ++i) {
 			const Position& position = (*positions)[i];
 			Tile* tile = map.getTile(position);
 			if(!tile)
